Added TC2.4. Changed type to decimal to allow for much bigger numbers.

+++ src/server/ConsoleAmountRetriever.cs
@@ -4,9 +4,9 @@ namespace server
 {
     class ConsoleAmountRetriever
     {
-        public virtual double GetAmount()
+        public virtual decimal GetAmount()
         {
-            return double.Parse(Console.ReadLine());
+            return decimal.Parse(Console.ReadLine());
         }
     }
 }
